+++ RELEASES.md
@@ -2,6 +2,7 @@
 
 ## Version 2.0
 ### TODO
+* (PENDING) Fix utility memoryfile examples
 * (PENDING) Finalizing the IPC library - what is planned for SFv2.0.
 * Cross-platform support:
 	* (DONE) Linux